+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### 12.119.0
 
+- `[initsystem]` Added [launchd](https://www.launchd.info) support
 - `[support/services]` Added package for collecting services info
 
 ### 12.118.0